qapitrace: Stage 1 with logic for adding RENDER and STATE groups
Next stage: When collapsed, re-title group's parent call to
            indicate a RENDER or STATE group.

            Likewise with glPushDebugGroup; when collapsed,
            re-title with its "message" argument to describe
            the group

SUMMARY:
Currently qapitrace allows an expandable grouping of calls within
a Frame that are called between glPushDebugGroup/glPopDebugGroup

This push/pop mechanism has been expanded to use another two set
of groupings based on the categories:

   RENDER
   STATE

such that initially, instead of a Frame expanding to all the calls
within the Frame, it now expands only to groups of calls within the
Frame which in turn are expandable to the individual calls

This maintains the glPush/PopDebugGroup groups but adds groups of
calls delimited by a RENDER call or glPushDebugGroup call.

DETAILS:
If a RENDER call is encountered, then it, plus all the previous calls
up to the parent call, is constituted as a RENDER group.

If a glPushDebugGroup is encountered, then the previous calls, up
to their parent are consolidated into a STATE group (with the
following exceptions) and the glPushDebugGroup starts a new set of
grouped calls.

Exceptions:
* If glPushDebugGroup is the first call in a Frame
* If the parent is a glPushDebugGroup

For an unpaired glPushDebugGroup call within a frame, it will
continue collecting calls until terminated by a END_FRAME call.

For an unpaired glPopDebugGroup call within a frame, it will close
the preceding group as STATE, add itself to the parent as an
ungrouped call and then initiate the next call as start of a new
group. This will make unpaired glPopDebugGroups stand out.
(The other option is to keep glPopDebugGroup as an additional
member call of the existing STATE or RENDER group.)

TESTING:
Additional testing shows that this grouping doesn't work if one
of the calls that starts a group gets filtered out - then the
whole group of calls disappears.