qapitrace: Add a filter object member to TraceLoader
Add ApiTraceFilter object to TraceLoader class to check if the
call is going to be filtered out before making it start of a
group; otherwise the whole group gets filtered out

CAVEAT:
Note that this doesn't allow for changing options. I.e. the tree
doesn't get rebuilt so that previously filtered items that would
have been the head of a group are now all individual calls

Also fixed standalone glPopDebugGroup so if they're unpaired, they
don't start a new group as a regular call.

Also make FRAME_END call child of frame, not child of last group
(if in group)